requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asqlite @ git+https://github.com/Rapptz/asqlite@fcd8ce0672562e440f99eb4b7a56eba16f6abf4e
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
braceexpand==0.1.7
charset-normalizer==3.3.2
click==8.1.7
discord.py==2.3.2
dnspython==2.4.2
frozenlist==1.4.0
idna==3.4
import-expression==1.1.4
jishaku==2.5.2
motor==3.3.1
multidict==6.0.4
pymongo==4.6.0
python-dotenv==1.0.0
six==1.16.0
yarl==1.9.2